add initial serde support for relevant types
Serde is an important part of the Rust ecosystem, and we want to ensure users
can securely serialize and deserialize types easily and securely. The Serialize
and Deserialize implementations here delegate to constant-time operations
wherever possible.

The phrase "relevant types" above means any type that we expect a user may want
to send outside the application, e.g. to a database or across the network. For
now, this includes:

    - pwhash::PasswordHash
    - hash::Digest
    - auth::Tag
    - kdf::Salt

By not implementing serde for other types, we're discouraging users from using
them insecurely.